requirements.txt
fastapi[all]
anthropic
requests
uvicorn
python-dotenv